% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_write_to_log}
\alias{map_write_to_log}
\title{Map helper function for use with log creation}
\usage{
map_write_to_log(check_d, log_df, log_file)
}
\arguments{
\item{check_d}{A description of the `check` to be added to the list}

\item{log_df}{A dataframe corresponding to the log entries for the asociated `check`}

\item{log_file}{The log file to be used in the process of binding all together.}
}
\value{
This function will process each of the entries in teh corresponding collectioan of `checks`.
}
\description{
Supplementary function to assist with the main map used to create the logs. This is an internal function and there should be no need to access it directly.
}
\examples{






}
