% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_equal}
\alias{test_equal}
\title{Test to check whether two values are equal.}
\usage{
test_equal(descr, source_value, test_value, rounding_digits = 1, ...)
}
\arguments{
\item{descr}{A text string to explain what this test is about}

\item{source_value}{The one end of the comaprison - This is the source/origin you are comapring to}

\item{test_value}{The other end of the comparison - This is the tested value}

\item{rounding_digits}{This is the number of rounding up digits to be used when comparing. Default is 1 digit}
}
\value{
All  test functions will assign their output to a global variable so as to enable reusing them in multiple `tests` and `checks`
}
\description{
The way this function works it will allow for some digits to be rounded off in case
numbers with minor differences
}
\examples{

## You can choose to use the function on its own but is best suited for use within a `check` and
## this within a `QA_that` environment

test_equal(descr = "check that 12 matches with 12.2 using rounding up", 12,12.4, rounding_digits = 0)

}
