% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_identical}
\alias{test_identical}
\title{Test to check whether two values are identical}
\usage{
test_identical(descr, source_value, test_value, ...)
}
\arguments{
\item{descr}{A text string to explain what this test is about}

\item{source_value}{The one end of the comaprison - This is the source/origin you are comapring to}

\item{test_value}{The other end of the comparison - This is the tested value}
}
\value{
All  test functions will assign their output to a global variable so as to enable reusing them in multiple `tests` and `checks`
}
\description{
This function will test for strict equality between a source and a test value
}
\examples{

## You can choose to use the function on its own but is best suited for use within a `check` and
## this within a `QA_that` environment

test_identical(descr = "test description text", 12,12.1)

}
