% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_match}
\alias{test_match}
\title{Test to check whether two character vectors are identical}
\usage{
test_match(descr, source_value, test_value, ...)
}
\arguments{
\item{descr}{A text string to explain what this test is about}

\item{source_value}{The one end of the comaprison - This is the source/origin you are comapring to}

\item{test_value}{The other end of the comparison - This is the tested value}
}
\value{
All  test functions will assign their output to a global variable so as to enable reusing them in multiple `tests` and `checks`.This is a
             binary TRUE or FALSE result for each test.
}
\description{
This function is meant to run in the context of a `check` and not on its own. You can run it in stand alone mode to test the output but
the corresponding global container will be deleted when a `check` is ran.
}
\examples{

## You can choose to use the function on its own but is best suited for use within a `check` and
## this within a `QA_that` environment

output44 = check_that(check_desc = "numeric matching test 1", {test_equal(descr = "test description text", 11,11)
 test_equal(descr = "test description text", 12,12)
 test_identical(descr = "test description text", 12,12.1)
 test_match(descr = "checking that two test strings match", "DFDF","dfdf")})
}
