% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_match}
\alias{test_match}
\title{Test to match whether a string matches a given source value}
\usage{
test_match(descr, source_value, test_value, ...)
}
\arguments{
\item{descr}{A text string to explain what this test is about}

\item{source_value}{The one end of the comaprison - This is the source/origin you are comapring to}

\item{test_value}{The other end of the comparison - This is the tested value}
}
\value{
All  test functions will assign their output to a global variable so as to enable reusing them in multiple `tests` and `checks`
}
\description{
This functioasn will test to see if two strings are identical
}
\examples{

## You can choose to use the function on its own but is best suited for use within a `check` and
## this within a `QA_that` environment

test_match(descr = "checking that two test strings match", "DFDF","dfdf")

}
