% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{use_QA}
\alias{use_QA}
\title{Creating folder structure for QA}
\usage{
use_QA(QAname)
}
\arguments{
\item{QAname}{The name of the first QA file to be created. This will be an `R` file and it is only required to supply just the name without the suffix `.R` of the file.}
}
\value{
The function will return a folder structure in the root of the main workign directory and a first QA file to get things started. In addition, a `QA_logs` folder
will also be created to store the corresponding logs from the QA process.
}
\description{
This will create a folder structure to use with the `QA.that` package. The important thing to note here is that the function needs to be used from
the root working directory otherwise it will be created in a different place than the one intended. The folde rname is set to be `QA_that` and is supposed to be static
so as to enable picking it up from otehr functions for later porcessing.
}
\examples{

useQA("QA that table matches table B")





}
